# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ifandonlyif API
  version: "1.0"
paths:
  /code:
    post:
      consumes:
      - application/json
      description: register a new user
      parameters:
      - description: walletAddress
        in: body
        name: walletAddress
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: code
      tags:
      - code
  /gasinfo:
    get:
      description: get 24 hours gas prices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: gasinfo
      tags:
      - gasinfo
  /health:
    get:
      consumes:
      - '*/*'
      description: get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Show the status of server.
      tags:
      - root
  /login:
    post:
      consumes:
      - application/json
      description: login
      parameters:
      - description: walletAddress
        in: body
        name: walletAddress
        required: true
        schema:
          type: string
      - description: nonce
        in: body
        name: nonce
        required: true
        schema:
          type: string
      - description: signature
        in: body
        name: signature
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: login
      tags:
      - login
schemes:
- http
swagger: "2.0"